requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.5.0
psycopg2-binary>=2.9.9

# SaaS Platform additions
openpyxl>=3.1.2
reportlab>=4.0.0
psutil>=5.9.0
bcrypt>=4.0.0
